% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{sim_data}
\usage{
sim_data(S, K, lambda, phi, theta)
}
\arguments{
\item{S}{Number of sites}

\item{K}{Number of visits}

\item{lambda}{Expected abundance}

\item{phi}{Proportion of individuals that are marked}

\item{theta}{Visitation rate}
}
\description{
Simulate counts and capture-recapture data
}
